requirements.txt
django==4.2.7
django-environ==0.11.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
django-jazzmin==2.6.0

argon2-cffi==23.1.0
djoser==2.2.2 
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
social-auth-app-django==5.4.0
social-auth-core==4.5.2

Pillow==10.1.0 
django-storages==1.14.2
django-ckeditor==6.7.0
psycopg2==2.9.9 
python-slugify==8.0.4 